backend/requirements.txt
# Django Core
Django
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Authentication
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
djoser==2.2.0

# Database
psycopg2-binary==2.9.9
pymongo==4.5.0

# WebSocket Support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Async Tasks
celery==5.3.4
redis==5.0.1

# File Handling
Pillow==10.1.0
python-magic==0.4.27

# Rich Text Editor
django-ckeditor==6.7.0

# API Documentation
drf-yasg==1.21.7

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Development
django-extensions==3.2.3
ipython==8.17.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring
sentry-sdk==1.38.0

# Utils
python-dateutil==2.8.2
pytz==2023.3

# Chat functionality
markdown2==2.4.10
bleach==6.1.0
django-cleanup==8.0.0